+++ python/common/hotkeys_window.py
@@ -18,7 +18,7 @@ class HotkeysWindow(QMainWindow):
         hotkeys: RTSHotkeys,
         game_icon: str,
         mouse_image: str,
-        conifguration_folder: str,
+        configuration_folder: str,
         panel_settings: RTSHotkeysConfigurationLayout,
         timer_flag: bool = False,
     ):
@@ -30,7 +30,7 @@ class HotkeysWindow(QMainWindow):
         hotkeys                 Hotkeys current definition.
         game_icon               Icon of the game.
         mouse_image             Image for the mouse.
-        conifguration_folder    Folder with the configuration files (settings and build orders).
+        configuration_folder    Folder with the configuration files (settings and build orders).
         panel_settings          Settings for the panel layout.
         timer_flag              True to add the timer hotkeys.
         """
@@ -127,7 +127,7 @@ class HotkeysWindow(QMainWindow):
         self.folder_button.setStyleSheet(self.style_button)
         self.folder_button.adjustSize()
         self.folder_button.move(self.border_size, y_buttons)
-        self.folder_button.clicked.connect(lambda: subprocess.run(['explorer', conifguration_folder]))
+        self.folder_button.clicked.connect(lambda: subprocess.run(['explorer', configuration_folder]))
         self.folder_button.show()
         first_column_max_width = max(first_column_max_width, widget_x_end(self.folder_button))