+++ Changelog.md
@@ -2,6 +2,8 @@
 * Web
     * Auto-resize is the default.
     * Shorter tooltips for overlay window.
+* Python
+    * Add back window to paste JSON code.
 
 # [2.13.0] - 2026.06.10
 * Web